% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MEView.read.file.header}
\alias{MEView.read.file.header}
\title{Read data optained with MEView}
\usage{
See MEView-40 user guide.
}
\arguments{
\item{file.name:}{Name of file to be read}
}
\value{
Alist with data and meta data in separate dataframes.
}
\description{
Read data optained with MEView
}
\author{
Claus E. Andersen
}

